Fix refreshing behavior, create funtion for fetching zip
